fast-api-pydantic-sqlalchemy/requirements.txt
fastapi
pydantic
sqlalchemy
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]